requirements.txt
Flask>=2.2.3
Flask_Cors>=3.0.10
matplotlib>=3.7.1
mido>=1.2.10
numpy>=1.23.5
python-dotenv>=1.0.0
tensorflow>=2.12.0
gunicorn==20.1.0
scipy==1.10.1